Fixed previous merge and max file issues
Deleted OSRM old data with plans to not host osrm files on github in the future. Working on Docker file as wel as pather to setup deploying the server as a whole within a docker. Was able to run the osrm docker and generate paths given waypoints using electric wheelchair profiles.